+++ Package.swift
@@ -1,0 +1,22 @@
+// swift-tools-version:5.9
+import PackageDescription
+
+let package = Package(
+    name: "QvpKit",
+    platforms: [.iOS(.v15), .macOS(.v12)],
+    products: [
+        .library(name: "QvpKit", targets: ["QvpKit"]),
+    ],
+    targets: [
+        .binaryTarget(
+            name: "QvpEngine",
+            url: "https://github.com/quran-ws/quran-engine/releases/download/v0.2.1/QvpEngine.xcframework.zip",
+            checksum: "9490ec567790874295a3cb611472f4185dbb43ca76bf360e53d13fec460f29de"
+        ),
+        .target(
+            name: "QvpKit",
+            dependencies: ["QvpEngine"],
+            path: "packages/ios/QvpKit/Sources/QvpKit"
+        ),
+    ]
+)